+++ src/index.html
@@ -23,11 +23,6 @@
 
     <include src="./partials/modal-film.html"></include>
     <include src="./partials/modal-team.html"></include>
-    <a href="#top" class="scroll-up-btn scroll-up-btn--hidden">
-      <svg class="scroll-up-btn__icon" width="30" height="30">
-        <use href="./images/sprite.svg#arrow-up"></use>
-      </svg>
-    </a>
     <!-- <script type="module" src="./js/firebase/firebase.js"></script> -->
     <!-- <script type="module" src="./js/renderTrendigMovies/renderTrendingPage.js"></script> -->
     <!-- <script type="module" src="./js/modalFilm/renderModalFilm.js"></script> -->
