+++ src/index.js
@@ -3,4 +3,3 @@ import './js/renderTrendigMovies/renderTrendingPage';
 import './js/modalFilm/renderModalFilm';
 import './js/filmsByName/fetchFilmsByName';
 // import './js/createTrailer/createTrailer';
-import './js/hideScrollUpBtn';
