+++ src/library.html
@@ -24,11 +24,6 @@
 
     <include src="./partials/modal-film.html"></include>
     <include src="./partials/modal-team.html"></include>
-    <a href="#top" class="scroll-up-btn scroll-up-btn--hidden">
-      <svg class="scroll-up-btn__icon" width="30" height="30">
-        <use href="./images/sprite.svg#arrow-up"></use>
-      </svg>
-    </a>
     <script type="module" src="./library.js"></script>
     <!-- <script type="module" src="./js/modalFilm/renderModalFIlmLibr.js"></script> -->
   </body>
