+++ src/library.js
@@ -1,3 +1,2 @@
 import './js/renderQueueWatchedFilms/renderQueueWachedFilms';
 import './js/modalFilm/renderModalFilm';
-import './js/hideScrollUpBtn';
