+++ src/sass/index.scss
@@ -6,7 +6,6 @@
 
 @import './components/modal-team';
 @import './components/modal-film';
-@import './components/scroll-button';
 
 @import './layout/home-header';
 @import './layout/libr-header';
